pyproject.toml
[tool.poetry]
name = "awesomedimgrayflatassembler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
opentelemetry-api = "^0.14b0"
opentelemetry-sdk = "^0.14b0"
opentelemetry-instrumentation = "^0.14b0"
opentelemetry-opentracing-shim = "^0.14b0"
opentelemetry-exporter-jaeger = "^0.14b0"
redis = "^3.5.3"
redis-opentracing = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
